Fixed some long-standing compiler warnings in the rtree code.  Seems to still work properly.

+++ src/rtree/rect.c
@@ -369,7 +369,7 @@ int Xastir_RTreeContained(struct Rect *R, struct Rect *S)
 {
 	register struct Rect *r = R, *s = S;
 	register int i, j, result;
-	assert((int)r && (int)s);
+	assert(r && s);
 
  	// undefined rect is contained in any other
 	//
